// src/modules/questions/question.enum.ts
export enum QuestionTopic {
  SLEEP = 'sleep',
  ENERGY = 'energy',
  ANXIETY = 'anxiety',
  DIZZINESS = 'dizziness',
  NAUSEA = 'nausea',
  TUNNEL_VISION = 'tunnel_vision',
  HEARING_SENSITIVE = 'unpleasant_hearing_sensitive',
  TREMBLING_BODY = 'trembling_body',
  SEPERATE_THOUNGHT = 'seperation_thought',
  MIND_QUIETER = 'mind_quieter',
  AWARENESS_BODY = 'awareness_body',
}

export enum QuestionType {
  YES_NO = 'yes_no',
  RATING = 'rating',
}